完善博客，问答，课程，社区，社区文章，资源删除功能

+++ monkey-question/src/main/java/com/monkey/monkeyquestion/service/impl/UserHomeQuestionServiceImpl.java
@@ -2,6 +2,12 @@ package com.monkey.monkeyquestion.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.monkey.monkeyUtils.constants.CommonEnum;
+import com.monkey.monkeyUtils.constants.MessageEnum;
+import com.monkey.monkeyUtils.constants.ReportCommentEnum;
+import com.monkey.monkeyUtils.constants.ReportContentEnum;
+import com.monkey.monkeyUtils.mapper.*;
+import com.monkey.monkeyUtils.pojo.*;
 import com.monkey.monkeyUtils.result.R;
 import com.monkey.monkeyquestion.feign.QuestionToSearchFeignService;
 import com.monkey.monkeyquestion.mapper.*;
@@ -36,6 +42,18 @@ public class UserHomeQuestionServiceImpl implements UserHomeQuestionService {
     private QuestionReplyCommentMapper questionReplyCommentMapper;
     @Resource
     private QuestionToSearchFeignService questionToSearchFeignService;
+    @Resource
+    private MessageLikeMapper messageLikeMapper;
+    @Resource
+    private MessageCollectMapper messageCollectMapper;
+    @Resource
+    private ReportContentMapper reportContentMapper;
+    @Resource
+    private MessageCommentReplyMapper messageCommentReplyMapper;
+    @Resource
+    private ReportCommentMapper reportCommentMapper;
+    @Resource
+    private CollectContentConnectMapper collectContentConnectMapper;
 
     /**
      * 通过用户id查询用户发布问答
@@ -127,12 +145,64 @@ public class UserHomeQuestionServiceImpl implements UserHomeQuestionService {
             // 删除问答回复表
             questionReplyMapper.deleteBatchIds(questionReplyIdList);
 
-            // 删除提问回复评论表
-            LambdaQueryWrapper<QuestionReplyComment> questionReplyCommentLambdaQueryWrapper = new LambdaQueryWrapper<>();
-            questionReplyCommentLambdaQueryWrapper.in(QuestionReplyComment::getQuestionReplyId, questionReplyIdList);
-            questionReplyCommentMapper.delete(questionReplyCommentLambdaQueryWrapper);
+            // 得到提问回复表id
+            LambdaQueryWrapper<QuestionReplyComment> replyCommentLambdaQueryWrapper = new LambdaQueryWrapper<>();
+            replyCommentLambdaQueryWrapper.in(QuestionReplyComment::getQuestionReplyId, questionReplyIdList);
+            replyCommentLambdaQueryWrapper.select(QuestionReplyComment::getId);
+            List<Object> commentIdList = questionReplyCommentMapper.selectObjs(replyCommentLambdaQueryWrapper);
+            if (commentIdList != null && commentIdList.size() > 0) {
+                // 删除提问回复评论表
+                LambdaQueryWrapper<QuestionReplyComment> questionReplyCommentLambdaQueryWrapper = new LambdaQueryWrapper<>();
+                questionReplyCommentLambdaQueryWrapper.in(QuestionReplyComment::getQuestionReplyId, questionReplyIdList);
+                questionReplyCommentMapper.delete(questionReplyCommentLambdaQueryWrapper);
+
+                // 删除消息回复评论表
+                LambdaQueryWrapper<MessageCommentReply> messageCommentReplyLambdaQueryWrapper = new LambdaQueryWrapper<>();
+                messageCommentReplyLambdaQueryWrapper.eq(MessageCommentReply::getAssociationId, questionId);
+                messageCommentReplyLambdaQueryWrapper.in(MessageCommentReply::getCommentId, commentIdList);
+                messageCommentReplyLambdaQueryWrapper.eq(MessageCommentReply::getType, ReportCommentEnum.QUESTION_REPORT.getCode());
+                messageCommentReplyMapper.delete(messageCommentReplyLambdaQueryWrapper);
+
+                // 删除举报评论表
+                LambdaQueryWrapper<ReportComment> reportCommentLambdaQueryWrapper = new LambdaQueryWrapper<>();
+                reportCommentLambdaQueryWrapper.eq(ReportComment::getType, ReportCommentEnum.QUESTION_REPORT.getCode());
+                reportCommentLambdaQueryWrapper.in(ReportComment::getAssociateId, commentIdList);
+                reportCommentMapper.delete(reportCommentLambdaQueryWrapper);
+
+                // 删除消息评论点赞表
+                LambdaQueryWrapper<MessageLike> messageCommentLikeLambdaQueryWrapper = new LambdaQueryWrapper<>();
+                messageCommentLikeLambdaQueryWrapper.eq(MessageLike::getAssociationId, questionId);
+                messageCommentLikeLambdaQueryWrapper.eq(MessageLike::getType, MessageEnum.QUESTION_MESSAGE.getCode());
+                messageCommentLikeLambdaQueryWrapper.in(MessageLike::getCommentId, commentIdList);
+                messageCommentLikeLambdaQueryWrapper.eq(MessageLike::getIsComment, CommonEnum.MESSAGE_LIKE_IS_COMMENT.getCode());
+                messageLikeMapper.delete(messageCommentLikeLambdaQueryWrapper);
+            }
         }
 
+        // 删除消息点赞表
+        LambdaQueryWrapper<MessageLike> messageLikeLambdaQueryWrapper = new LambdaQueryWrapper<>();
+        messageLikeLambdaQueryWrapper.eq(MessageLike::getAssociationId, questionId);
+        messageLikeLambdaQueryWrapper.eq(MessageLike::getType, MessageEnum.QUESTION_MESSAGE.getCode());
+        messageLikeLambdaQueryWrapper.eq(MessageLike::getIsComment, CommonEnum.MESSAGE_LIKE_IS_CONTENT.getCode());
+        messageLikeMapper.delete(messageLikeLambdaQueryWrapper);
+
+        // 删除消息收藏表
+        LambdaQueryWrapper<MessageCollect> messageCollectLambdaQueryWrapper = new LambdaQueryWrapper<>();
+        messageCollectLambdaQueryWrapper.eq(MessageCollect::getAssociationId, questionId);
+        messageCollectLambdaQueryWrapper.eq(MessageCollect::getType, MessageEnum.QUESTION_MESSAGE.getCode());
+        messageCollectMapper.delete(messageCollectLambdaQueryWrapper);
+
+        // 删除举报内容表
+        LambdaQueryWrapper<ReportContent> reportContentLambdaQueryWrapper = new LambdaQueryWrapper<>();
+        reportContentLambdaQueryWrapper.eq(ReportContent::getAssociateId, questionId);
+        reportContentLambdaQueryWrapper.eq(ReportContent::getType, ReportContentEnum.QUESTION_REPORT.getCode());
+        reportContentMapper.delete(reportContentLambdaQueryWrapper);
+
+        // 删除收藏目录关系表
+        LambdaQueryWrapper<CollectContentConnect> collectContentConnectLambdaQueryWrapper = new LambdaQueryWrapper<>();
+        collectContentConnectLambdaQueryWrapper.eq(CollectContentConnect::getType, CommonEnum.COLLECT_QUESTION.getCode())
+                .eq(CollectContentConnect::getAssociateId, questionId);
+        collectContentConnectMapper.delete(collectContentConnectLambdaQueryWrapper);
 
         // 删除elasticsearch问答表
         questionToSearchFeignService.deleteQuestion(String.valueOf(questionId));
